+++ lib/td_dq/rules/rules.ex
@@ -424,6 +424,21 @@ defmodule TdDq.Rules do
     query |> Repo.all()
   end
 
+  def get_rule_implementation_results(implementation_key) do
+    RuleResult
+    |> where([r], r.implementation_key == ^implementation_key)
+    |> order_by(desc: :date)
+    |> Repo.all()
+  end
+
+  def get_last_rule_result(implementation_key) do
+    RuleResult
+    |> where([r], r.implementation_key == ^implementation_key)
+    |> order_by(desc: :date)
+    |> limit(1)
+    |> Repo.one()
+  end
+
   @doc """
   Returns the list of rule_implementations.