+++ lib/td_dq_web/swagger_definitions.ex
@@ -110,6 +110,25 @@ defmodule TdDqWeb.SwaggerDefinitions do
           type(:array)
           items(Schema.ref(:Rule))
         end,
+      RulesExecuteRequest:
+        swagger_schema do
+          properties do
+            search_params(:object, "Search params")
+          end
+        end,
+      RulesIDs:
+        swagger_schema do
+          title("Rules IDs")
+          description("Rules IDs to execute")
+          type(:array)
+          items(%{type: :integer})
+        end,
+      RulesExecuteResponse:
+        swagger_schema do
+          properties do
+            data(Schema.ref(:RulesIDs))
+          end
+        end,
       RuleResponse:
         swagger_schema do
           properties do